services/dice-roller/pyproject.toml
[tool.poetry]
name = "dice-roller"
version = "0.1.0"
description = ""
authors = ["Johan Hellmark <johan.hellmark50@gmial.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opentelemetry-distro = "^0.41b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
